--- a/e2e/cloudwalk-contracts/integration/test/leader-follower-change.test.ts
+++ b/e2e/cloudwalk-contracts/integration/test/leader-follower-change.test.ts
@@ -62,8 +62,7 @@ describe("Leader & Follower change integration test", function () {
 
     it("Validate new Follower health after change", async function () {
         updateProviderUrl("stratus");
-        await sendAndGetFullResponse("stratus_changeToLeader", []);
-        await new Promise((resolve) => setTimeout(resolve, 10000));
+        await new Promise((resolve) => setTimeout(resolve, 20000));
         const response = await sendWithRetry("stratus_health", []);
         expect(response).to.equal(true);
     });
@@ -90,8 +89,7 @@ describe("Leader & Follower change integration test", function () {
 
     it("Validate new Leader health after change", async function () {
         updateProviderUrl("stratus-follower");
-        await sendAndGetFullResponse("stratus_changeToLeader", []);
-        await new Promise((resolve) => setTimeout(resolve, 10000));
+        await new Promise((resolve) => setTimeout(resolve, 20000));
         const response = await sendWithRetry("stratus_health", []);
         expect(response).to.equal(true);
     });