### **PR Type**
Enhancement, Tests


___

### **Description**
- Removed redundant `stratus_changeToLeader` calls in the "Validate new Follower health after change" and "Validate new Leader health after change" test cases
- Optimized test execution time by reducing the sleep duration from 10 seconds to 5 seconds
- Improved test efficiency without compromising the validation of leader and follower health after changes


___



### **Changes walkthrough** 📝
<table><thead><tr><th></th><th align="left">Relevant files</th></tr></thead><tbody><tr><td><strong>Enhancement</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>leader-follower-change.test.ts</strong><dd><code>Optimize leader-follower change tests</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

e2e/cloudwalk-contracts/integration/test/leader-follower-change.test.ts

<li>Removed unnecessary <code>stratus_changeToLeader</code> calls in two test cases<br> <li> Reduced sleep time from 10000ms to 5000ms<br>


</details>


  </td>
  <td><a href="https://github.com/cloudwalk/stratus/pull/1695/files#diff-7236d4b676b75ced96eb321337290326efe4e91a261edc4e83b688c4da09dd7c">+0/-2</a>&nbsp; &nbsp; &nbsp; </td>

</tr>                    
</table></td></tr></tr></tbody></table>

___

> 💡 **PR-Agent usage**:
>Comment `/help` on the PR to get a list of all available PR-Agent tools and their descriptions